requirements.txt
streamlit>=1.28.0
google-generativeai
Pillow>=10.0.0
requests>=2.30.0
json5>=0.9.0
pyttsx3>=2.90
pygame>=2.5.0
gtts>=2.4.0
